changed ushort to uint

+++ src/GameServer/RemoteView/Character/UpdateCharacterStatsPlugIn.cs
@@ -66,7 +66,7 @@ public class UpdateCharacterStatsPlugIn : IUpdateCharacterStatsPlugIn
             (ushort)this._player.SelectedCharacter.UsedNegFruitPoints,
             this._player.SelectedCharacter.GetMaximumFruitPoints(),
             (byte)this._player.SelectedCharacter.InventoryExtensions,
-            (ushort)this._player.Attributes[Stats.Resets])
+            (uint)this._player.Attributes[Stats.Resets])
             .ConfigureAwait(false);
 
         if (this._player.SelectedCharacter.CharacterClass!.IsMasterClass)
